Ajout du job d'ajout des personnes
